transaction-service/app/requirements.txt
Flask
kafka-python
psychopg2-binary
